% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sCVPP.R
\name{approx.bf.p}
\alias{approx.bf.p}
\title{This function computes posterior probabilities for a variant to be causal using Wakefields asymptotic Bayes Factors.}
\usage{
approx.bf.p(p, f, type = c("QUANT", "CC"), N, s, pi_i = 1e-04)
}
\arguments{
\item{p}{vector of p.values for $i$ univariate tests of association for the trait of interest with a set of variants}

\item{f}{a vector of length $i$ of minor allele frequencies in controls matching the order of p}

\item{type}{a scalar taking the  of the proportion of cases for each test.}

\item{N}{a scalar or vector (of length $i$) of the total number of samples included in association test}

\item{s}{a scalar or vector (of length $i$) of the proportion of cases for each test.}

\item{pi_i}{a scalar of the prior probability that a given variant is causal (default=1e-4)}
}
\value{
a vector of prior probabilities.
}
\description{
This function computes posterior probabilities for a variant to be causal using Wakefields asymptotic Bayes Factors.
}
