% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{csnps.gr}
\alias{csnps.gr}
\title{csnps.gr
Coding SNPs for chromosome 22.}
\format{GRanges object with 53002 ranges and 2 metadata columns.}
\usage{
csnps.gr
}
\description{
csnps.gr
Coding SNPs for chromosome 22.
}
\keyword{datasets}
