% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{digest.gr}
\alias{digest.gr}
\title{digest.gr
HindIII restriction digest for chromosome 22.}
\format{GRanges object with 7582 ranges and 1 metadata column.}
\usage{
digest.gr
}
\description{
digest.gr
HindIII restriction digest for chromosome 22.
}
\keyword{datasets}
