% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{feature.sets}
\alias{feature.sets}
\title{feature.sets
Feature sets from Javierre et al pcHi-C dataset for chromosome 22.}
\format{list of 18 data.table objects}
\usage{
feature.sets
}
\description{
feature.sets
Feature sets from Javierre et al pcHi-C dataset for chromosome 22.
}
\keyword{datasets}
