% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{fix_chr}
\alias{fix_chr}
\title{This function fixes chromosomes so they are numeric}
\usage{
fix_chr(chr, f)
}
\arguments{
\item{chr}{a vector of strings representing chromosomes}

\item{f}{a scalar representing a file name (so user can understand which file is problematic)}
}
\value{
a vector of reformatted chromosomes
}
\description{
This function fixes chromosomes so they are numeric
}
