% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pchic.R
\name{load_capture_design}
\alias{load_capture_design}
\title{This loads a capture design file with columns fragid and ensg. fragid should match digest file.}
\usage{
load_capture_design(f)
}
\arguments{
\item{f}{a scalar representing a file name of design file}
}
\value{
data.table object representing capture design
}
\description{
This loads a capture design file with columns fragid and ensg. fragid should match digest file.
}
