% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pchic.R
\name{load_digest}
\alias{load_digest}
\title{This loads a restriction digest file with columns chr,start,end and fragid}
\usage{
load_digest(f)
}
\arguments{
\item{f}{a scalar representing a file name of digest file}
}
\value{
GRanges object representing restriction digest
}
\description{
This loads a restriction digest file with columns chr,start,end and fragid
}
