% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{load_gwas}
\alias{load_gwas}
\title{This function loads and checks set of univariate P values for a given trait and adds posterior probabilities for each SNP to be causal
using Wakefield's aBF}
\usage{
load_gwas(f, maf.DT, ld.gr, n.cases, n.controls, plink = FALSE)
}
\arguments{
\item{f}{character scalar representing GWAS file to load, tab delimited with three columns chr,pos and p}

\item{maf.DT}{a data.table of reference MAFs created by load_ref_maf}

\item{ld.gr}{a GenomicRanges object of LD regions created by load_ld_regions}

\item{n.cases}{a scalar for the number of GWAS cases - this should equal sample size in the case of a quantitative trait}

\item{n.controls}{a scalar for the number of GWAS controls - this should be omitted in the case of a quantitative trait}

\item{plink}{a boolean as to whether input format is PLINK rather than the cannonical type described above.}
}
\value{
a data.table of annotated GWAS variants suitable for COGS analysis
}
\description{
This function loads and checks set of univariate P values for a given trait and adds posterior probabilities for each SNP to be causal
using Wakefield's aBF
}
