% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{load_ld_regions}
\alias{load_ld_regions}
\title{This function checks and loads and checks a set of non overlapping  LD regions expects three column tab delim file with colnames chr,start,end}
\usage{
load_ld_regions(f)
}
\arguments{
\item{f}{character scalar representing filename containing maf data expects three columns wtih names chr, start and end.}
}
\value{
GenomicRanges object of ld blocks/regions
}
\description{
This function checks and loads and checks a set of non overlapping  LD regions expects three column tab delim file with colnames chr,start,end
}
