% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{load_ref_maf}
\alias{load_ref_maf}
\title{This function checks and loads a reference set of maf expects three columns tab delim file with colnames  chr,pos and maf}
\usage{
load_ref_maf(f, min.maf = 0.05)
}
\arguments{
\item{f}{character scalar representing filename containing maf data expects three columns wtih names chr, pos and maf.}

\item{min.maf}{scalar numeric representing minor allele threshold cut off.}
}
\value{
data.table object of minor allele freq.
}
\description{
This function checks and loads a reference set of maf expects three columns tab delim file with colnames  chr,pos and maf
}
