% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pchic.R
\name{make_csnps}
\alias{make_csnps}
\title{This function loads in a coding SNPs matched to a given ensembl id, columns are chr,pos and ensg}
\usage{
make_csnps(f, keep.ensg)
}
\arguments{
\item{f}{a scalar representing file name of coding SNP file}

\item{keep.ensg}{a vector of ensembl id's to include. If ommited then all ensembl ids regardless of further filtering are included}
}
\value{
GRanges object of coding SNPs
}
\description{
This function loads in a coding SNPs matched to a given ensembl id, columns are chr,pos and ensg
}
