% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pchic.R
\name{make_pchic}
\alias{make_pchic}
\title{This function computes loads in a set of promoter capture Hi-C files (pcHi-C). This is a tab delimited file that contains the following mandatory columns
\enumerate{
\item ensg - Ensembl gene id for gene
\item name - Gene name
\item biotype - Gene biotype from ensembl e.g. protein_coding
\item strand - Gene strand
\item baitChr - Bait (A captured restriction fragment) chromosome
\item baitStart - Bait fragment start
\item baitEnd - Bait fragment end
\item baitID - Bait fragment ID should match ID in digest file
\item baitName - Free text of what is captured - deprecated.
\item oeChr - 'Other end' (Promoter interacting restriction fragment) chromosome
\item oeStart - 'Other end' fragment start
\item oeEnd - 'Other end' fragment end
\item oeID - 'Other end' fragment ID should match ID in digest file
\item oeName - Free text of what is interacting - deprecated.
}
Additional columns contain CHiCAGO scores for one or more columns.}
\usage{
make_pchic(f, chic_thresh = 5, biotype.filter = "protein_coding",
  remove.trans.contacts = TRUE)
}
\arguments{
\item{f}{a scalar representing file name of pcHi-C file}

\item{chic_thresh}{scalar representing CHiCAGO threshold for calling a significant interaction. Default is set to 5.}

\item{biotype.filter}{scalar/vector of characters representing a biotype filter. Default is set to 'protein_coding'}

\item{remove.trans.contacts}{a boolean on whether to remove interaction between chromosomes. Default is set to TRUE.}
}
\value{
list of GenomicRanges representing regions interacting with genes for each tissue type.
}
\description{
This function computes loads in a set of promoter capture Hi-C files (pcHi-C). This is a tab delimited file that contains the following mandatory columns
\enumerate{
\item ensg - Ensembl gene id for gene
\item name - Gene name
\item biotype - Gene biotype from ensembl e.g. protein_coding
\item strand - Gene strand
\item baitChr - Bait (A captured restriction fragment) chromosome
\item baitStart - Bait fragment start
\item baitEnd - Bait fragment end
\item baitID - Bait fragment ID should match ID in digest file
\item baitName - Free text of what is captured - deprecated.
\item oeChr - 'Other end' (Promoter interacting restriction fragment) chromosome
\item oeStart - 'Other end' fragment start
\item oeEnd - 'Other end' fragment end
\item oeID - 'Other end' fragment ID should match ID in digest file
\item oeName - Free text of what is interacting - deprecated.
}
Additional columns contain CHiCAGO scores for one or more columns.
}
