% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pchic.R
\name{make_vprom}
\alias{make_vprom}
\title{This function computes 'Virtual' promoter regions. These are baited fragments and the fragments either side.}
\usage{
make_vprom(f.digest, f.design, keep.ensg)
}
\arguments{
\item{f.digest}{a scalar representing a file name of digest file}

\item{f.design}{a scalar representing a file name of design file}

\item{keep.ensg}{a vector of ensembl id's to include. If ommited then all ensembl ids regardless of further filtering are included}
}
\value{
GRanges object of 'Virtual' promoter regions
}
\description{
This function computes 'Virtual' promoter regions. These are baited fragments and the fragments either side.
}
