% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{t1d.DT}
\alias{t1d.DT}
\title{t1d.DT
GWAS data from Cooper et al. for Type 1 diabetes for chromosome 22.}
\format{data.table with 68300 rows and 7 variables
\describe{
  \item{pid}{unique id obtained from chr and position}
  \item{chr}{chromosome}
  \item{pos}{position GRCh37}
  \item{p}{univariate p value}
  \item{maf}{minor allele freq from maf.DT}
  \item{ld}{ld block from ld.gr}
  \item{ppi}{posterior probability for variant to be causal}
}}
\usage{
t1d.DT
}
\description{
t1d.DT
GWAS data from Cooper et al. for Type 1 diabetes for chromosome 22.
}
\keyword{datasets}
