% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sCVPP.R
\name{Var.data.cc}
\alias{Var.data.cc}
\title{helper function to compute shrinkage factor for a case/control trait}
\usage{
Var.data.cc(f, N, s)
}
\arguments{
\item{f}{a vector of length $i$ of minor allele frequencies in controls matching the order of p}

\item{N}{a scalar or vector (of length $i$) of the total number of samples included in association test}

\item{s}{a scalar or vector (of length $i$) of the proportion of cases for each test.}
}
\value{
a vector of shrinkage factors
}
\description{
helper function to compute shrinkage factor for a case/control trait
}
